House password
Input: A password as a string (Unicode for python 2.7).
Output: Is the password safe or not as a boolean or any data type that
can be converted and processed as a boolean. In the results you will see
the converted results.

Example:
checkio('A1213pokl') == False
checkio('bAse730onE') == True
checkio('asasasasasasasaas') == False
checkio('QWERTYqwerty') == False
checkio('123456123456') == False
checkio('QwErTy911poqqqq') == True